requirements.txt
chromadb
langchain-huggingface
llama-parse
streamlit
google-generativeai